Scripts/requirements.txt
google-api-python-client==2.48.0